Moved UDEBase and helper functions to utils and made UDEBase's hash function universal (hopefully)
